Update studio-configure-vm-task-unx.adoc
The original document doesn't provide guidance on the proper way for finding your JDK 1.8 installation, which many users don't know. If they've installed multiple JDK's this can become a big headache. The document also doesn't provide enough guidance for finding the AnypointStudio.ini file, which is hidden away under an unexpected path, especially if users happen to try to follow [this old support article](https://help.mulesoft.com/s/article/Increasing-JVM-memory-heap-in-Anypoint-Studio-to-avoid-OutOfMemory-issues) (maybe it was written for a version of Studio or MacOSX that was very different, but this path doesn't exist for me or my colleagues running  AnyPoint Studio 7.3 and OSX 10.13 - 10.15.